lab5/zoomcamp/pipeline/assets/ingestion/requirements.txt
# Python dependencies for the trips ingestion asset
pandas==2.2.0
requests==2.31.0
python-dateutil==2.8.2
pyarrow==15.0.0

